Add files via upload
This is a Diwali sales Data of a company. In this project I did cleaned and manipulated the data. 
I did the eploratory data analysis and found out some amazing insights from the Data using Numpy, Pandas, matplotlib and seaborn libraries.
See the Conclusion in the last for the Insights.
Special Thanks to Rishab Mishra. (instagram.com/rishabhnmishra)